Crate refinery_core

Re-exports

pub use crate::error::Error;

pub use crate::traits::async::AsyncMigrate;

pub use crate::traits::sync::Migrate;

pub use mysql;

Modules

config

error

traits

Structs

Migration

Represents a schema migration to be run on the database, this struct is used by the embed_migrations! macro to gather migration files and shouldn’t be needed by the user

Report

Struct that represents the report of the migration cycle, a Report instance is returned by the Runner::run and Runner::run_async methods via Result<Report, Error>, on case of an Error during a migration, you can access the Report with Error.report

Runner

Struct that represents the entrypoint to run the migrations, an instance of this struct is returned by the embed_migrations! macro. Runner should not need to be instantiated manually

Enums

MigrationType

enum containing the migration types used to search for migrations either just .sql files or both .sql and .rs

Target

An enum set that represents the target version up to which refinery should migrate, it is used by Runner

Functions

find_migration_files

find migrations on file system recursively across directories given a location and MigrationType